fix: wrong return from get and list
After a change in the API, GetMigration and ListMigration were improperly returning values that only supported outbound migrations.

The tricky bit here was that MigrationState could hold either an Inbound or an Outbound in the Migration field.

To deal with this I used a json.RawMessage to hold the field's contents temporarily. Alternative approaches resulted in a field full of horrifying nested maps no developer was ever meant to see.

Once we have the RawMessage we can properly unmarshal it and return things. The end user should have an easier time as they can call GetMigrationType() and then type assert to the correct value.