rpsr: add rpsr package, Redpanda Schema registry
This is a client wrapper that extends the franz-go
sr client to handle the upcoming Schema Registry
ACL endpoints.

It includes a helper ACLBuilder which allow an
ergonomic way of handling ACL creations with
optional defaults. The interface and errors are
similar to franz-go acl package with the difference
that here we do not assume any default value and
must be specified.