rpadmin: refactor to support admin API v2.
This is an internal refactor that splits our
internal request handlers into different reusable
functions so we can use our AdminAPI struct as a
client for the new Admin API v2.

This way, upgrading to the new version will not
require spinning up a new instance, but instead
reuse the same one you already have, including
TLS, and Auth Management.

The only breaking change we are doing at the moment
is the signature of our Opt type, now you can use
it to modify the adminapi client properties and
not just the retryable client.

Our long-term goal is to cleanup the multiple ways
we have to create a client in this package and
opt into (pun intended) to Opts.